add secure

+++ apps/documentation-front/action/article.action.ts
@@ -1,11 +1,23 @@
 "use server";
 
-import { config } from "@/config";
+import { config } from "@/config/config";
+import { auth } from "@clerk/nextjs/server";
 import { Article } from "@repo/ui/models/article.models";
 
 export const getArticles = async () => {
   try {
-    const res = await fetch(`${config.apiUrl}/articles`);
+    const { getToken } = await auth();
+    const token = await getToken();
+    if (!token) {
+      console.error("Unauthorized");
+      return [];
+    }
+
+    const res = await fetch(`${config.API_URL}/articles`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
     return res.json() as Promise<Article[]>;
   } catch (error) {
     console.error(error);
@@ -15,7 +27,18 @@ export const getArticles = async () => {
 
 export const getArticle = async (id: string) => {
   try {
-    const res = await fetch(`${config.apiUrl}/articles/${id}`);
+    const { getToken } = await auth();
+    const token = await getToken();
+    if (!token) {
+      console.error("Unauthorized");
+      return [];
+    }
+
+    const res = await fetch(`${config.API_URL}/articles/${id}`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
     return res.json() as Promise<Article>;
   } catch (error) {
     console.error(error);
@@ -27,10 +50,18 @@ export const createArticle = async (
   article: Omit<Article, "id" | "createdAt" | "updatedAt">
 ) => {
   try {
-    const res = await fetch(`${config.apiUrl}/articles`, {
+    const { getToken } = await auth();
+    const token = await getToken();
+    if (!token) {
+      console.error("Unauthorized");
+      return [];
+    }
+
+    const res = await fetch(`${config.API_URL}/articles`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
       body: JSON.stringify(article),
     });
@@ -46,10 +77,18 @@ export const updateArticle = async (
   article: Partial<Omit<Article, "id" | "createdAt" | "updatedAt">>
 ) => {
   try {
-    const res = await fetch(`${config.apiUrl}/articles/${id}`, {
+    const { getToken } = await auth();
+    const token = await getToken();
+    if (!token) {
+      console.error("Unauthorized");
+      return [];
+    }
+
+    const res = await fetch(`${config.API_URL}/articles/${id}`, {
       method: "PUT",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
       body: JSON.stringify(article),
     });
@@ -62,8 +101,18 @@ export const updateArticle = async (
 
 export const deleteArticle = async (id: string) => {
   try {
-    const res = await fetch(`${config.apiUrl}/articles/${id}`, {
+    const { getToken } = await auth();
+    const token = await getToken();
+    if (!token) {
+      console.error("Unauthorized");
+      return [];
+    }
+
+    const res = await fetch(`${config.API_URL}/articles/${id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
     });
     return res.ok;
   } catch (error) {
